Added Javascript functions
Was able to get the mobile formatting down to something simpler. I was also able to add Script.JS

I wrote out all the functions, and rendered the web page. Currently it has functions to remove the div class "main", and now just need to type out the HTML code for the Project page, and Bio page.